+++ Models/Cards/CardAttachment.cs
@@ -6,9 +6,9 @@ public class CardAttachment
 {
     public int Id { get; set; }
     
-    public string Url { get; set; }
+    public string? Url { get; set; }
     
-    public string Description { get; set; }
+    public string? Description { get; set; }
     
     [JsonIgnore]
     public EducationalCard EducationalCard { get; set; }
