+++ Program.cs
@@ -23,7 +23,11 @@ builder.Services.AddMySQLServer<ApplicationDbContext>(connectionString);
 builder.Services.AddScoped<IPasswordHasher<User>, PasswordHasher<User>>();
 builder.Services.AddScoped<IUserService, UserService>();
 builder.Services.AddScoped<IAuthenticationService, AuthenticationService>();
-builder.Services.AddScoped<ILessonService, LessonService>();
+builder.Services.AddScoped<ILessonManagementService, LessonManagementService>();
+builder.Services.AddScoped<ICardManagementService, CardManagementService>();
+builder.Services.AddScoped<IQuestionService, QuestionService>();
+builder.Services.AddScoped<IEducationalDataService, EducationalDataService>();
+builder.Services.AddScoped<ITaskService, TaskService>();
 builder.Services.AddScoped<ITestService, TestService>();
 builder.Services.AddScoped<ILessonTrackerService, LessonTrackerService>();
 builder.Services.AddEndpointsApiExplorer();
